Move ImageLoader from graphics folder to utils.
